Refactor utility functions in to src/backends/einsum_util.ts
